+++ NEWS.md
@@ -8,6 +8,10 @@
 
 * Increased datetime precision to the second, for `pin_versions()` and related
   functions (#642, @tomsing1).
+  
+* Pass the dots from `pin_write()` through to `s3_upload_file()` and 
+  `s3_uploade_yaml()` to support S3 tagging, encryption options, etc for 
+  pins (#648, #652, @fh-mthomson).  
 
 # pins 1.0.2